refactor(backup): extract cron schedule helpers to shared utility
Move ScheduleType, ScheduleState, DAY_NAMES, parseCronString, and
buildCronString from MetadataConfigSection into frontend/src/utils/cronSchedule.ts
for reusability and separation of concerns.
